requirements.txt
ftfy==6.2.0
jittor==1.3.8.5
numpy==1.24.4
opencv-python==4.6.0.66
Pillow==8.3.2
regex==2024.5.15
torch==1.9.0+cu111
tqdm==4.66.4